requirements-api.txt
Flask==2.0.1
Flask-CORS==3.0.10
Werkzeug==2.0.1
requests==2.31.0
cloudinary==1.36.0
qrcode[pil]==7.4.2
Pillow==10.0.1
supabase==2.0.0
numpy==1.24.3
librosa==0.10.1
ImageHash==4.3.1
pytesseract==0.3.10
gunicorn==21.2.0
